requirements.txt
# Core Framework
fastapi==0.124.0
uvicorn[standard]==0.37.0
pydantic==2.10.6
pydantic-settings==2.12.0

# Data Fetching & Analysis
yfinance==0.2.66
pandas==2.3.0
numpy==2.3.0

# LLM Integration
google-genai==1.54.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.38
alembic==1.16.5

# Payment Gateway
midtransclient==1.4.2

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.20
httpx==0.28.1
mangum==0.19.0

# Chart Generation
matplotlib==3.10.0

# Telegram Bot
python-telegram-bot==21.8